rip out all notion stuff, unused packages, etc
