+++ next/app/globals.css
@@ -34,7 +34,6 @@
     --ring: 215 20.2% 65.1%;
 
     --radius: 0.5rem;
-    --notion-link-underline-thickness: 2px !important;
     --fg-color-underline: rgba(55, 53, 47, 1) !important;
   }
 
@@ -83,286 +82,4 @@
 }
 
 @layer components {
-}
-
-.notion-collection-card {
-  @apply border-none shadow-none hover:bg-accent transition-all duration-200 rounded-lg p-2 pb-0 !important;
-}
-
-.notion-collection-card-cover {
-  @apply rounded-md border-none !important;
-}
-
-.notion-collection-card-body {
-  @apply flex-row flex-wrap items-center !important;
-}
-
-.notion-collection-card-body > *:first-child {
-  @apply flex-grow w-full !important;
-}
-
-.notion-collection-card-body > *:last-child {
-  @apply ml-4 text-sm flex-grow  justify-end flex !important;
-}
-
-.notion-collection-card-body .notion-property-text {
-  @apply font-normal text-sm py-1 px-2 bg-secondary rounded-full !important;
-}
-
-.notion-page-icon-inline {
-  display: none;
-}
-
-.notion-page-link {
-  height: auto !important;
-}
-
-.notion-collection-page-properties .notion-property-text {
-  @apply text-xl italic text-secondary-foreground/70 mb-4 block !important;
-}
-
-.notion-collection-page-properties .notion-property-date {
-  @apply text-lg !important;
-}
-
-.notion-list-body {
-  border-top: none !important;
-  /* gap: 8px; */
-}
-
-.notion-asset-wrapper {
-  position: relative;
-  @apply mb-4 !important;
-}
-
-.notion-asset-wrapper > *:after {
-  content: "";
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  box-shadow: 0 0 0 1px hsl(var(--border)) inset;
-  width: 100%;
-  height: 100%;
-  pointer-events: none;
-}
-
-.notion-asset-caption {
-  @apply text-center p-2 md:p-4 border-t border-border !important;
-}
-
-.notion-title {
-  @apply leading-normal text-5xl md:text-6xl font-light tracking-tight font-headline mb-2 !important;
-  line-height: 1.2 !important;
-}
-
-.notion-text,
-.notion-list,
-.notion-quote {
-  @apply text-sm sm:text-base md:text-lg mb-4 !important;
-}
-
-.notion-collection-column-title {
-  width: auto !important;
-  @apply h-0 mr-2 text-sm sm:text-base md:text-lg hidden !important;
-}
-
-.notion-collection-column-title > svg {
-  display: none;
-}
-
-.notion-property-multi_select-item {
-  @apply bg-transparent w-auto p-0 underline !important;
-}
-
-.notion-property {
-  @apply text-sm sm:text-base md:text-lg !important;
-}
-
-.notion-property-date {
-  @apply text-sm text-secondary-foreground/50 !important;
-}
-
-.notion-collection-row-value {
-  @apply p-0 !important;
-}
-
-.notion-collection-row {
-  @apply pb-8 mb-8 border-b border-border !important;
-}
-
-.notion-collection-row-property .notion-link {
-  border-bottom: 1px solid !important;
-}
-
-.logotype {
-  @apply font-serif text-2xl;
-  background: linear-gradient(
-    90deg,
-    rgba(28, 127, 238, 1) 0%,
-    rgba(95, 21, 242, 1) 20%,
-    rgba(186, 12, 248, 1) 40%,
-    rgba(251, 7, 217, 1) 60%,
-    rgba(255, 0, 0, 1) 100%
-  );
-  @apply drop-shadow-lg bg-clip-text text-transparent hover:text-transparent transition-all duration-500 tracking-tight;
-}
-
-.notion-gallery-grid {
-  @apply border-none !important;
-}
-
-.notion-page-title-text {
-  @apply whitespace-normal font-light text-sm sm:text-base md:text-lg font-sans !important;
-}
-
-.notion-property-title {
-  @apply whitespace-nowrap text-sm sm:text-base md:text-lg font-sans font-normal !important;
-}
-
-.notion-list-item-title {
-  @apply max-w-full !important;
-}
-
-.notion-h-title {
-  @apply font-light font-serif !important;
-}
-
-.notion-hr {
-  @apply border mt-6 mb-8 !important;
-}
-
-.notion > * {
-  padding: 0 !important;
-}
-
-.notion-list-body {
-  overflow: auto !important;
-}
-
-.notion-list-item-body {
-  flex-shrink: 0;
-  overflow: auto !important;
-}
-
-.notion-list-item-body > *:first-child {
-  margin: 0;
-}
-
-.notion-collection {
-  max-width: calc(100vw - 2rem);
-  width: calc(100% + 24px);
-  overflow: auto !important;
-  margin-left: -12px !important;
-  margin-right: -12px !important;
-}
-
-.notion-list-item {
-  @apply gap-1 py-2 w-full max-w-none overflow-auto flex flex-col sm:flex-row items-start sm:items-center !important;
-}
-
-.notion-list-item.notion-page-link {
-  --fg-color-underline: transparent;
-  --notion-link-underline-thickness: 0px !important;
-  transition: all 300ms ease-in-out;
-}
-
-@media (min-width: 768px) {
-  .notion-list-item {
-    padding: 8px 12px !important;
-  }
-
-  .notion-list-item.notion-page-link > :last-child:after {
-    content: "";
-    display: block;
-    width: 32px;
-    height: 32px;
-    margin-left: 8px;
-    background-image: url('data:image/svg+xml,<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="18.1348" height="14.8145"><g><rect height="14.8145" opacity="0" width="18.1348" x="0" y="0"/><path d="M18.1348 7.40234C18.1348 7.1582 18.0371 6.93359 17.8418 6.74805L11.3672 0.292969C11.1523 0.078125 10.9473 0 10.7129 0C10.2344 0 9.86328 0.351562 9.86328 0.839844C9.86328 1.07422 9.94141 1.29883 10.0977 1.45508L12.2852 3.68164L16.1621 7.2168L16.3574 6.72852L13.2129 6.5332L0.859375 6.5332C0.351562 6.5332 0 6.89453 0 7.40234C0 7.91016 0.351562 8.27148 0.859375 8.27148L13.2129 8.27148L16.3574 8.07617L16.1621 7.59766L12.2852 11.123L10.0977 13.3496C9.94141 13.4961 9.86328 13.7305 9.86328 13.9648C9.86328 14.4531 10.2344 14.8047 10.7129 14.8047C10.9473 14.8047 11.1523 14.7168 11.3477 14.5312L17.8418 8.05664C18.0371 7.87109 18.1348 7.64648 18.1348 7.40234Z" fill="%23000" fill-opacity="0.85"/></g></svg>');
-    background-repeat: no-repeat;
-    background-position: 100%;
-    opacity: 0;
-    transition: all 200ms ease-in-out;
-    width: 32px;
-    will-change: width, opacity;
-    transform: scale-x(0);
-  }
-
-  .notion-list-item.notion-page-link:hover {
-    background: rgba(0, 0, 0, 0.03) !important;
-    border-radius: 8px;
-  }
-
-  .notion-list-item.notion-page-link:hover > :last-child:after {
-    opacity: 1;
-    width: 32px;
-    transform: scale-x(1);
-    background-position: center;
-  }
-}
-
-.tape {
-  background-color: hsla(0, 0%, 100%, 0.2);
-  box-shadow: inset 0 0 1em 0.5em hsla(0, 0%, 100%, 0.1);
-  height: 2em;
-  position: absolute;
-  margin: auto;
-  left: 0;
-  right: 0;
-  top: -5%;
-  width: 40%;
-  -webkit-filter: drop-shadow(0 1px 1px hsla(0, 0%, 0%, 0.3));
-}
-.tape:after,
-.tape:before {
-  background-size: 0.4em 0.4em;
-  bottom: 0;
-  content: "";
-  position: absolute;
-  top: 0;
-  width: 0.2em;
-}
-.tape:after {
-  background-image: linear-gradient(
-      45deg,
-      transparent 50%,
-      hsla(0, 0%, 100%, 0.3) 50%
-    ),
-    linear-gradient(-45deg, transparent 50%, hsla(0, 0%, 100%, 0.3) 50%);
-  background-position: 0 100%;
-  left: -0.2em;
-}
-.tape:before {
-  background-image: linear-gradient(
-      135deg,
-      transparent 50%,
-      hsla(0, 0%, 100%, 0.3) 50%
-    ),
-    linear-gradient(-135deg, transparent 50%, hsla(0, 0%, 100%, 0.3) 50%);
-  background-position: 100% 100%;
-  right: -0.2em;
-}
-
-.postcard {
-  border: 8px solid;
-  border-image: url("/stickers/border.svg");
-  border-image-slice: 6;
-  border-image-repeat: repeat fill;
-}
-
-.bits-page .notion-title {
-  text-align: center;
-}
-
-.bits-page .notion-page-content-inner > .notion-text:first-of-type {
-  text-align: center;
-}
-
-.notion-inline-code {
-  @apply bg-secondary rounded-sm px-1 text-primary !important;
-  --notion-link-underline-thickness: 1px;
-}
-
-.notion-asset-wrapper.notion-asset-wrapper-video > div:has(video) {
-  height: auto !important;
 }
