+++ next/app/src/components/mdx/createHeading.tsx
@@ -1,12 +1,22 @@
 import React from "react"
 
 import { cn } from "../../lib/utils"
-import { slugify } from "./mdx"
 
 interface HeadingProps {
   children?: any
   className?: string
   text?: string
+}
+
+export function slugify(str) {
+  return str
+    .toString()
+    .toLowerCase()
+    .trim() // Remove whitespace from both ends of a string
+    .replace(/\s+/g, "-") // Replace spaces with -
+    .replace(/&/g, "-and-") // Replace & with 'and'
+    .replace(/[^\w\-]+/g, "") // Remove all non-word characters except for -
+    .replace(/\-\-+/g, "-") // Replace multiple - with single -
 }
 
 export function createHeading(level) {
@@ -20,7 +30,9 @@ export function createHeading(level) {
     "text-base",
   ]
   return ({ children, className, text }: HeadingProps) => {
-    let slug = slugify(text || children?.props?.children || "")
+    const child =
+      typeof children === "string" ? children : children?.props?.children
+    let slug = slugify(text || child || "")
     return React.createElement(
       `h${level}`,
       { id: slug, className: cn(baseTextStyle, sizes[level - 1], className) },
