+++ next/app/src/components/mdx/mdx.tsx
@@ -152,17 +152,6 @@ function Video({ className, ...props }) {
   return <video className={cn("mb-4 mx-auto", className)} {...props} />
 }
 
-export function slugify(str) {
-  return str
-    .toString()
-    .toLowerCase()
-    .trim() // Remove whitespace from both ends of a string
-    .replace(/\s+/g, "-") // Replace spaces with -
-    .replace(/&/g, "-and-") // Replace & with 'and'
-    .replace(/[^\w\-]+/g, "") // Remove all non-word characters except for -
-    .replace(/\-\-+/g, "-") // Replace multiple - with single -
-}
-
 let components = {
   h1: createHeading(1),
   h2: createHeading(2),
