--- next/app/src/components/notion/styles.module.css
@@ -1,88 +1,0 @@
-@keyframes spinner {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.container {
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: calc(min(2vmin, 24px));
-}
-
-.loadingIcon {
-  animation: spinner 0.6s linear infinite;
-  display: block;
-  width: 24px;
-  height: 24px;
-  color: currentColor;
-}
-
-.renderer {
-  --notion-font: var(--font-sans);
-  --fg-color: hsl(var(--primary));
-  --fg-color-0: hsl(var(--border));
-  --fg-color-1: hsl(var(--border));
-  --fg-color-2: hsl(var(--primary) / 0.7);
-  --fg-color-3: hsl(var(--primary) / 0.5);
-  --fg-color-4: #e2e2e2;
-  --fg-color-5: rgba(237, 237, 237, 0.024);
-  --fg-color-6: rgba(231, 231, 231, 0.8);
-  --fg-color-icon: var(--primary);
-
-  --bg-color: hsl(var(--background));
-  --bg-color-0: hsl(var(--background) / 0.6);
-  --bg-color-1: hsl(var(--background) / 0.4);
-  --bg-color-2: hsl(var(--background) / 0.2);
-
-  --select-color-0: rgb(46, 170, 220);
-  --select-color-1: rgba(45, 170, 219, 0.3);
-  --select-color-2: #d9eff8;
-
-  --notion-red: var(--red-200);
-  --notion-pink: rgb(173, 26, 114);
-  --notion-blue: rgb(11, 110, 153);
-  --notion-purple: rgb(105, 64, 165);
-  --notion-teal: rgb(77, 100, 97);
-  --notion-yellow: rgb(223, 171, 1);
-  --notion-orange: rgb(217, 115, 13);
-  --notion-brown: rgb(100, 71, 58);
-  --notion-gray: rgb(155, 154, 151);
-
-  --notion-red_background: #f2baba;
-  --notion-pink_background: #f7c6db;
-  --notion-blue_background: #c2d0f8;
-  --notion-purple_background: #d7c0fb;
-  --notion-teal_background: #b1f2ed;
-  --notion-yellow_background: #faefc3;
-  --notion-orange_background: #ffe0d4;
-  --notion-brown_background: #fadfca;
-  --notion-gray_background: #e0e0e0;
-
-  --notion-red_background_co: rgba(251, 228, 228, 0.3);
-  --notion-pink_background_co: rgba(244, 223, 235, 0.3);
-  --notion-blue_background_co: rgba(221, 235, 241, 0.3);
-  --notion-purple_background_co: rgba(234, 228, 242, 0.3);
-  --notion-teal_background_co: rgba(221, 237, 234, 0.3);
-  --notion-yellow_background_co: rgba(251, 243, 219, 0.3);
-  --notion-orange_background_co: rgba(250, 235, 221, 0.3);
-  --notion-brown_background_co: rgba(233, 229, 227, 0.3);
-  --notion-gray_background_co: rgba(235, 236, 237, 0.3);
-
-  --notion-item-blue: rgba(0, 120, 223, 0.2);
-  --notion-item-orange: rgba(245, 93, 0, 0.2);
-  --notion-item-green: rgba(0, 135, 107, 0.2);
-  --notion-item-pink: rgba(221, 0, 129, 0.2);
-  --notion-item-brown: rgba(140, 46, 0, 0.2);
-  --notion-item-red: rgba(255, 0, 26, 0.2);
-  --notion-item-yellow: rgba(233, 168, 0, 0.2);
-  --notion-item-default: rgba(206, 205, 202, 0.5);
-  --notion-item-purple: rgba(103, 36, 222, 0.2);
-  --notion-item-gray: rgba(155, 154, 151, 0.4);
-
-  --notion-max-width: 720px;
-  --notion-header-height: 45px;
-}
