--- next/app/src/types/notion-compat/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'notion-compat';
