--- next/app/talks/[slug]/page.tsx
@@ -1,5 +1,0 @@
-function Page({ params }: { params: { slug: string } }) {
-  return <div>Talk post: {params.slug}</div>
-}
-
-export default Page
