--- next/app/talks/page.tsx
@@ -1,5 +1,0 @@
-function Page() {
-  return <div>Talks</div>
-}
-
-export default Page
