+++ next/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "next-template",
-  "version": "0.0.2",
+  "name": "ryngonzalez.com",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -15,7 +15,6 @@
   },
   "dependencies": {
     "@codesandbox/sandpack-react": "^2.13.5",
-    "@notionhq/client": "^1.0.4",
     "@radix-ui/react-slot": "^1.0.2",
     "@tailwindcss/container-queries": "^0.1.1",
     "@uidotdev/usehooks": "^2.3.1",
@@ -24,25 +23,14 @@
     "class-variance-authority": "^0.4.0",
     "clsx": "^1.2.1",
     "framer-motion": "^10.12.16",
-    "got": "^13.0.0",
-    "lqip-modern": "^2.0.0",
     "lucide-react": "^0.105.0-alpha.4",
     "next": "^13.4.4",
     "next-mdx-remote": "^4.4.1",
     "next-themes": "^0.2.1",
-    "notion-client": "https://gitpkg.now.sh/ryngonzalez/react-notion-x/packages/notion-client?33a78ed3c185b07166e722ca031cef1df5607bde",
-    "notion-compat": "^6.16.0",
-    "notion-utils": "^6.16.0",
-    "p-map": "^6.0.0",
-    "p-memoize": "^7.1.1",
-    "prismjs": "^1.29.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.10",
-    "react-notion-x": "https://gitpkg.now.sh/ryngonzalez/react-notion-x/packages/react-notion-x?33a78ed3c185b07166e722ca031cef1df5607bde",
+    "react-loader-spinner": "^6.1.6",
     "react-tweet": "^3.2.0",
-    "react-tweet-embed": "^2.0.0",
-    "sharp": "^0.31.3",
     "sugar-high": "^0.6.0",
     "tailwind-merge": "^1.13.0",
     "tailwindcss-animate": "^1.0.5",
@@ -50,13 +38,10 @@
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^3.7.2",
-    "@types/lqip-modern": "^1.1.3",
     "@types/node": "^17.0.45",
-    "@types/prismjs": "^1.26.0",
     "@types/react": "^18.2.7",
     "@types/react-dom": "^18.2.4",
     "@typescript-eslint/parser": "^5.59.7",
-    "@vercel/blob": "^0.10.0",
     "autoprefixer": "^10.4.14",
     "dotenv": "^16.3.1",
     "eslint": "^8.41.0",
