--- next/pages-old/bits-old/[pageId].tsx
@@ -1,69 +1,0 @@
-import * as React from "react"
-import { GetStaticPaths, GetStaticPropsContext } from "next"
-import { ExtendedRecordMap } from "notion-types"
-
-import { NotionPage } from "@/app/src/components/notion/NotionPage"
-import {
-  isDev,
-  previewImagesEnabled,
-  rootBitsDomain,
-  rootBitsNotionPageId,
-} from "@/app/src/config/notion"
-import * as notion from "@/app/src/lib/notion"
-import {
-  ResolveNotionPageResult,
-  resolveNotionPage,
-} from "@/app/src/lib/resolve-notion-page"
-
-export const getStaticProps = async (context: GetStaticPropsContext) => {
-  const rawPageId = context?.params?.pageId as string
-  const props = (await resolveNotionPage(
-    "ryngonzalez.com",
-    rawPageId,
-    rootBitsNotionPageId,
-    rootBitsDomain
-  )) as ResolveNotionPageResult
-
-  return {
-    props,
-    // we will attempt to re-generate the page:
-    // - when a request comes in
-    // - at most once every 10 seconds
-    revalidate: 10,
-  }
-}
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  if (isDev) {
-    return {
-      paths: [],
-      fallback: true,
-    }
-  }
-
-  // This crawls all public pages starting from the given root page in order
-  // for next.js to pre-generate all pages via static site generation (SSG).
-  // This is a useful optimization but not necessary; you could just as easily
-  // set paths to an empty array to not pre-generate any pages at build time.
-  const siteMap = await notion.getSiteMap(rootBitsNotionPageId, rootBitsDomain)
-
-  const paths = Object.keys(siteMap.canonicalPageMap)
-    .map((pageId) => ({ params: { pageId: pageId } }))
-    .filter((path) => path)
-
-  return {
-    paths,
-    fallback: true,
-  }
-}
-
-export default function Page({ recordMap }: { recordMap: ExtendedRecordMap }) {
-  return (
-    <NotionPage
-      recordMap={recordMap}
-      rootDomain={rootBitsDomain}
-      rootPageId={rootBitsNotionPageId}
-      previewImagesEnabled={previewImagesEnabled}
-    />
-  )
-}
